+++ drivers/ps2/ps2.go
@@ -1,6 +1,6 @@
 package ps2
 
-import "github.com/icexin/eggos/sys"
+import "github.com/icexin/eggos/kernel/sys"
 
 const (
 	_CMD_PORT  = 0x64
