+++ kernel/panic.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/icexin/eggos/debug"
 	"github.com/icexin/eggos/drivers/uart"
-	"github.com/icexin/eggos/sys"
+	"github.com/icexin/eggos/kernel/sys"
 )
 
 var (
