+++ kernel/seg_amd64.go
@@ -3,7 +3,7 @@ package kernel
 import (
 	"unsafe"
 
-	"github.com/icexin/eggos/sys"
+	"github.com/icexin/eggos/kernel/sys"
 )
 
 const (
