+++ kernel/sys/sys_amd64.s
