+++ kernel/trap.go
@@ -6,8 +6,8 @@ import (
 	"github.com/icexin/eggos/debug"
 	"github.com/icexin/eggos/drivers/pic"
 	"github.com/icexin/eggos/kernel/isyscall"
+	"github.com/icexin/eggos/kernel/sys"
 	"github.com/icexin/eggos/kernel/trap"
-	"github.com/icexin/eggos/sys"
 )
 
 var (
